Rename component to `ProjectAccessControls`
