+++ web-admin/src/components/projects/ProjectHero.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import AccessControls from "../access-controls/AccessControls.svelte";
+  import ProjectAccessControls from "./ProjectAccessControls.svelte";
   import ProjectDeploymentStatus from "./ProjectDeploymentStatus.svelte";
   import ProjectGithubConnection from "./ProjectGithubConnection.svelte";
   import ShareProjectCta from "./ShareProjectCTA.svelte";
@@ -12,7 +12,7 @@
   <h1 class="text-sky-950 text-4xl font-light leading-10">
     {project}
   </h1>
-  <AccessControls {organization} {project}>
+  <ProjectAccessControls {organization} {project}>
     <svelte:fragment slot="manage-project">
       <div
         class="flex flex-wrap gap-x-16 gap-y-8 items-start mt-8 mb-12 rounded w-fit"
@@ -22,7 +22,7 @@
         <ShareProjectCta {organization} {project} />
       </div>
     </svelte:fragment>
-  </AccessControls>
+  </ProjectAccessControls>
   <span class="text-gray-500 text-base font-normal leading-normal"
     >Check out your dashboards below.</span
   >
